requirements.txt
# Core dependencies
flask==2.0.1
gunicorn==20.1.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.10.0
python-dateutil==2.8.2
webdriver-manager==3.8.6

# Google Cloud dependencies
google-cloud-storage==2.9.0
google-cloud-secretmanager==2.16.1
google-cloud-firestore==2.11.1
google-cloud-bigquery==3.10.0
firebase-admin==6.1.0
google-cloud-logging==3.5.0

# Google Maps API
googlemaps==4.10.0

# OpenRouter for analysis
openai==1.3.0

# Misc utilities
retrying==1.3.4
validators==0.22.0